Taking into account PR Review:  - Reuse memory.go approach to grow: use slice backed array and let go manage memory allocation through append  - Add additional cap test to be sure all validator will act the same if stack exceed MaxInt32 depth
+ Add dedicated stack unit tests (heavily copied from memory_test.go and only for Push() + ensureCapacity)

+++ execution/config.go
@@ -15,8 +15,8 @@ const (
 
 type ExecutionConfig struct {
 	CallStackMaxDepth        uint64
-	DataStackInitialCapacity int
-	DataStackMaxDepth        int
+	DataStackInitialCapacity uint64
+	DataStackMaxDepth        uint64
 	VMOptions                []VMOption `json:",omitempty" toml:",omitempty"`
 }
